REFACTOR: Move ParsingException into its own file and some small refactoring
